move data type classes into utils to fix circular deps, WORKING and MODULARIZED
