+++ app/modules/core/model/mstlstm.py
@@ -6,7 +6,8 @@ from operator import itemgetter
 from dynet import ParameterCollection, AdamTrainer, LSTMBuilder, tanh, logistic, rectify, cmult, \
     SimpleRNNBuilder, concatenate, np, renew_cg, esum
 
-from .data_types import ConllEntry
+# from .data_types import ConllEntry
+from .utils import ConllEntry
 from .utils import read_conll
 from .decoder import parse_proj
